+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add additional security check editing php files in finder
 - Fix base64 encoding data not in latin char range
 - Add internal bucket storage container feature
+- Add /content/tree/{model} api endpoint
 - Update vendor libs
 
 ## 2.3.9 (2023-02-13)